Move CometScanRule into own file
